backend/requirements.txt
flask
pyflakes
jedi
python-dotenv
PyMySQL
pycryptodome
google-cloud-speech
scipy
noisereduce
pydub
wave
numpy
PySoundFile
python-Levenshtein